+++ lib/angular2/shared/orm/base.ejs
@@ -9,7 +9,7 @@ import { createIO } from './io';
 import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
 
-import { applyFilter } from './filter';
+import { applyFilter, toArray } from './filter';
 
 import { LoopBackFilter } from '../models';
 
@@ -23,17 +23,7 @@ export class OrmBase<T> {
 
       return applyFilter(
         this.store.select(this.model.getModelName() + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -44,17 +34,7 @@ export class OrmBase<T> {
 
       return applyFilter(
         this.store.select(this.model.getModelName() + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
         , filter, this.store, this.model);
     }
   }
@@ -100,17 +80,7 @@ export class OrmBase<T> {
 
       return applyFilter(
         this.store.select(this.model.getModelName() + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -123,17 +93,7 @@ export class OrmBase<T> {
 
       return applyFilter(
         this.store.select(this.model.getModelName() + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
         , newFilter, this.store, this.model).map((data: any[]) => {
           return data[0];
         });
