+++ lib/angular2/shared/orm/filter.ejs
@@ -158,3 +158,19 @@ function isPlainObject(obj: any): boolean {
   return (typeof obj === 'object') && (obj !== null) &&
     (obj.constructor === Object);
 }
+
+export function toArray(state: any): any[] {
+  const entities = [];
+  
+  for (let key in state.entities) {
+    if (state.entities.hasOwnProperty(key)) {
+      entities.push(state.entities[key]);
+    }
+  }
+
+  return entities;
+}
+
+export function filterById(state: any[], id: any, keyTo: string): any[] {
+  return state.filter((item: any) => item[keyTo] === id);
+}
