+++ lib/angular2/shared/orm/models/model.ejs
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
 
 import { OrmBase } from '../base';
-import { applyFilter } from '../filter';
+import { applyFilter, toArray, filterById } from '../filter';
 
 import * as models from '../../models';
 import { <%- modelName %>, LoopBackFilter } from '../../models';
@@ -47,17 +47,8 @@ public <%- normalizeMethodName(methodName) %>(<%- buildMethodParams(model, metho
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -68,17 +59,8 @@ public <%- normalizeMethodName(methodName) %>(<%- buildMethodParams(model, metho
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.model]);
     }
     <% } else { %>
@@ -86,7 +68,8 @@ public <%- normalizeMethodName(methodName) %>(<%- buildMethodParams(model, metho
 
     return applyFilter(
       this.store.select(this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.model + 's')
-        .map((state: any) => state.entities)
+        .map(toArray)
+        .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.keyTo]))
       , filter, this.store, models[this.model.getModelDefinition().relations.<%- methodName.match(/^__get__(\w+)$/)[1] %>.model]);
     <% } %>
   }
