+++ tests/ng2web/src/app/shared/sdk/orm/models/Account.ts
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
 
 import { OrmBase } from '../base';
-import { applyFilter } from '../filter';
+import { applyFilter, toArray, filterById } from '../filter';
 
 import * as models from '../../models';
 import { Account, LoopBackFilter } from '../../models';
@@ -136,17 +136,8 @@ export class OrmAccount extends OrmBase<Account> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.accessTokens.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.accessTokens.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -157,17 +148,8 @@ export class OrmAccount extends OrmBase<Account> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.accessTokens.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.accessTokens.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.accessTokens.model]);
     }
     
@@ -190,17 +172,8 @@ export class OrmAccount extends OrmBase<Account> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.rooms.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.rooms.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -211,17 +184,8 @@ export class OrmAccount extends OrmBase<Account> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.rooms.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.rooms.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.rooms.model]);
     }
     
@@ -244,17 +208,8 @@ export class OrmAccount extends OrmBase<Account> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.administrations.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.administrations.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -265,17 +220,8 @@ export class OrmAccount extends OrmBase<Account> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.administrations.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.administrations.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.administrations.model]);
     }
