+++ tests/ng2web/src/app/shared/sdk/orm/models/Category.ts
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
 
 import { OrmBase } from '../base';
-import { applyFilter } from '../filter';
+import { applyFilter, toArray, filterById } from '../filter';
 
 import * as models from '../../models';
 import { Category, LoopBackFilter } from '../../models';
@@ -68,17 +68,8 @@ export class OrmCategory extends OrmBase<Category> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.rooms.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.rooms.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -89,17 +80,8 @@ export class OrmCategory extends OrmBase<Category> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.rooms.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.rooms.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.rooms.model]);
     }
