+++ tests/ng2web/src/app/shared/sdk/orm/models/Message.ts
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
 
 import { OrmBase } from '../base';
-import { applyFilter } from '../filter';
+import { applyFilter, toArray, filterById } from '../filter';
 
 import * as models from '../../models';
 import { Message, LoopBackFilter } from '../../models';
@@ -134,17 +134,8 @@ export class OrmMessage extends OrmBase<Message> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.likes.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.likes.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -155,17 +146,8 @@ export class OrmMessage extends OrmBase<Message> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.likes.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.likes.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.likes.model]);
     }
     
@@ -188,17 +170,8 @@ export class OrmMessage extends OrmBase<Message> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.replies.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.replies.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -209,17 +182,8 @@ export class OrmMessage extends OrmBase<Message> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.replies.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.replies.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.replies.model]);
     }
