+++ tests/ng2web/src/app/shared/sdk/orm/models/User.ts
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
 
 import { OrmBase } from '../base';
-import { applyFilter } from '../filter';
+import { applyFilter, toArray, filterById } from '../filter';
 
 import * as models from '../../models';
 import { User, LoopBackFilter } from '../../models';
@@ -60,17 +60,8 @@ export class OrmUser extends OrmBase<User> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.accessTokens.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.accessTokens.keyTo]))
           .finally(() => {
             destroyStream$.next(1);
             destroyStream$.complete();
@@ -81,17 +72,8 @@ export class OrmUser extends OrmBase<User> {
 
       return applyFilter(
         this.store.select(this.model.getModelDefinition().relations.accessTokens.model + 's')
-          .map((state: any) => {
-            const entities = [];
-            
-            for (let key in state.entities) {
-              if (state.entities.hasOwnProperty(key)) {
-                entities.push(state.entities[key]);
-              }
-            }
-
-            return entities;
-          })
+          .map(toArray)
+          .map((state: any[]) => filterById(state, id, models[this.model.getModelDefinition().relations.accessTokens.keyTo]))
         , filter, this.store, models[this.model.getModelDefinition().relations.accessTokens.model]);
     }
